+++ source/app/service-providers/types/config-provider.d.ts
@@ -144,6 +144,7 @@ interface ConfigOptions {
   }
   system: {
     deleteOnFail: boolean
+    leaveAppRunning: boolean
     avoidNewTabs: boolean
     iframeWhitelist: string[]
   }