+++ source/common/vue/form/Form.vue
@@ -58,6 +58,8 @@
           v-bind:value="getModelValue(field.model)"
           v-bind:label="field.label"
           v-bind:name="field.model"
+          v-bind:disabled="field.disabled"
+          v-bind:info="field.info"
           v-on:input="$emit('input', field.model, $event)"
         ></CheckboxInput>
         <SwitchInput