+++ source/common/vue/form/elements/Checkbox.vue
@@ -1,16 +1,28 @@
 <template>
-  <div class="form-control cb-group">
-    <label class="checkbox">
-      <input
-        v-bind:id="fieldID"
-        type="checkbox" v-bind:name="name" value="yes"
-        v-bind:checked="value"
-        v-bind:disabled="disabled"
-        v-on:input="$emit('input', $event.target.checked)"
+  <div>
+    <div class="form-control cb-group">
+      <label class="checkbox"
+             v-bind:disabled="disabled"
       >
-      <span class="checkmark"></span>
-    </label>
-    <label v-if="label" v-bind:for="fieldID" v-html="label"></label>
+        <input
+          v-bind:id="fieldID"
+          type="checkbox"
+          v-bind:name="name"
+          value="yes"
+          v-bind:checked="value"
+          v-bind:disabled="disabled"
+          v-on:input="$emit('input', $event.target.checked)"
+        >
+        <span class="checkmark"></span>
+      </label>
+      <label v-if="label" v-bind:for="fieldID" v-bind:disabled="disabled"
+             v-html="label"
+      >
+      </label>
+    </div>
+    <div v-if="info" class="form-control info">
+      {{ info }}
+    </div>
   </div>
 </template>
 
@@ -47,6 +59,10 @@ export default {
     disabled: {
       type: Boolean,
       default: false
+    },
+    info: {
+      type: String,
+      default: ''
     }
   },
   computed: {
@@ -110,6 +126,26 @@ body {
         opacity: 1;
       }
     }
+
+    &[disabled] {
+      input:checked ~ .checkmark {
+        background-color: lightgrey;
+      }
+      input:checked ~ .checkmark {
+        border-color: rgb(90, 90, 90);
+      }
+    }
+  }
+
+  label {
+    &[disabled] {
+      color: grey;
+    }
+  }
+
+  div.info {
+    color: grey;
+    font-size: 80%;
   }
 }