+++ source/main/zettlr.ts
@@ -36,6 +36,7 @@ import { CodeFileDescriptor, CodeFileMeta, DirDescriptor, MDFileDescriptor, MDFi
 import broadcastIpcMessage from '../common/util/broadcast-ipc-message'
 
 export default class Zettlr {
+  isQuitting: boolean
   editFlag: boolean
   _openPaths: any
   _fsal: FSAL
@@ -49,6 +50,9 @@ export default class Zettlr {
     * @param {electron.app} parentApp The app object.
     */
   constructor () {
+    // Is the app quitting? True if quitting via menu, tray or keyboard shortcut.
+    // False if titlebar `x` close, and all other times.
+    this.isQuitting = false
     this.editFlag = false // Is the current opened file edited?
     this._openPaths = [] // Holds all currently opened paths.
     this.isShownFor = [] // Contains all files for which remote notifications are currently shown
@@ -62,6 +66,9 @@ export default class Zettlr {
     global.application = {
       runCommand: async (command: string, payload?: any) => {
         return await this.runCommand(command, payload)
+      },
+      isQuitting: () => {
+        return this.isQuitting
       },
       showLogViewer: () => {
         this._windowManager.showLogWindow()
@@ -80,6 +87,9 @@ export default class Zettlr {
       },
       showTagManager: () => {
         this._windowManager.showTagManager()
+      },
+      showAnyWindow: () => {
+        this._windowManager.showAnyWindow()
       },
       notifyChange: (msg: string) => {
         global.notify.normal(msg)
@@ -129,9 +139,11 @@ export default class Zettlr {
     // listen to close-events on the main window, we should be able to handle
     // this, if we ever switched to the auto updater.
     app.on('before-quit', (event) => {
+      this.isQuitting = true
       if (!this._documentManager.isClean()) {
         // Immediately prevent quitting ...
         event.preventDefault()
+        this.isQuitting = false
         // ... and ask the user if we should *really* quit.
         this._windowManager.askSaveChanges()
           .then(result => {