+++ webpack.main.config.js
@@ -20,7 +20,9 @@ module.exports = {
         { from: 'static/defaults', to: 'assets/defaults' },
         { from: 'static/export.tpl.htm', to: 'assets' },
         { from: 'static/template.revealjs.htm', to: 'assets' },
-        { from: 'static/revealjs-styles', to: 'assets/revealjs-styles' }
+        { from: 'static/revealjs-styles', to: 'assets/revealjs-styles' },
+        { from: 'resources/icons/icon.ico', to: 'assets/icons' },
+        { from: 'resources/icons/png', to: 'assets/icons/png' }
       ]
     })
   ],